Run a subset of release jobs on a custom runner
Some release jobs have to be executed on a custom runner due to GitHub forcing a
newer nodejs version: https://github.com/actions/checkout/issues/1809

This whole issue is a huge rabbit hole where you can get lost.
In short:

* The actions/checkout@4 versions require the nodejs 20 version
* Prebuilt nodejs 20 are linking against a newer GLIBC version
* Even though jobs are executed in a docker image, some commands are from
    the runner machine, (e.g. nodejs is from the ubuntu-latest runner).
* Downgrading to actions/checkout@3 worked for some time.

However, due to https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
it seems like we *have* to do something.
Thus, we are running jobs for older distro images on a customer runner, provided by maerwald.
In particular, when writing this message, the distro images that have be run on the custom runner are:

* debian:9
* ubuntu:18.04
* mint:19.3
* fedora:27
* centos:7